Changed codec write register function, data bit to reg addr

+++ SEAG_F/Core/Inc/Codec/CodecDriver.h
@@ -16,7 +16,7 @@ static const uint8_t codecAddr = 0x1A << 1; // Use 7-bit address
 
 void CodecInit(I2C_HandleTypeDef *hi2c1);
 void ResetCodec(I2C_HandleTypeDef *hi2c1);
-void WriteCodecRegister(I2C_HandleTypeDef *hi2c1, uint16_t reg, uint16_t pData);
+void WriteCodecRegister(I2C_HandleTypeDef *hi2c1, uint8_t reg, uint16_t pData);
 
 #endif /* INC_CODEC_CODECDRIVER_H_ */
